backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
groq==0.4.2
supabase==2.3.0
scikit-learn==1.4.1.post1
numpy==1.26.4
python-multipart==0.0.9
requests==2.31.0
PyPDF2==3.0.1
pydantic==2.6.1
pydantic_core==2.16.2
typing_extensions==4.9.0
